% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIRAL.R
\name{adjust.phases}
\alias{adjust.phases}
\title{Changes the inferred phases to obtain visually nicer plots using the fact that the phases live on a circular space where period-a=-a.}
\usage{
adjust.phases(realphi, infphi, period = 2 * pi)
}
\arguments{
\item{realphi}{Required vector of true phases.}

\item{infphi}{Required vector of inferred phases.}

\item{period}{Period of the space where the phases live. Default 2*pi}
}
\value{
Set of shifted inferred phases.
}
\description{
Changes the inferred phases to obtain visually nicer plots using the fact that the phases live on a circular space where period-a=-a.
}
\examples{
delta.phi(true.phi, inf.phi)
}
