% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIRAL.R
\name{CHIRAL}
\alias{CHIRAL}
\title{Infer the circular phase of gene expression samples}
\usage{
CHIRAL(
  E,
  iterations = 500,
  clockgenes = NULL,
  tau2 = NULL,
  u = NULL,
  sigma2 = NULL,
  TSM = TRUE,
  mean.centre.E = TRUE,
  q = 0.1,
  update.q = FALSE,
  pbar = TRUE,
  phi.start = NULL,
  standardize = FALSE,
  GTEx_names = FALSE
)
}
\arguments{
\item{E}{Required matrix of gene expression. Samples should be on the columns, genes on the rows. Rows should be named to perform the inference on a subset of the matrix.}

\item{iterations}{Number of maximum iterations, default 500.}

\item{clockgenes}{Set of clock genes (or relevant genes) to be used for the inference. Should be a of the form c("gene1", "gene2") and subset of the rownames, default NULL.}

\item{tau2}{Tau parameter for the prior on the gene coefficient, we suggest to leave the default. It is automatically calculated by the algorithm, can be changed if needed. Default NULL.}

\item{u}{u parameter for the prior on the gene means, we suggest to leave the default. It is automatically calculated by the algorithm, can be changed if needed. Default NULL.}

\item{sigma2}{standard deviation of the data points fro the prediction, we suggest to leave the default. It is automatically calculated by the algorithm, can be changed if needed. default NULL.}

\item{TSM}{Switches the two state model in the EM on and off, default TRUE.}

\item{mean.centre.E}{parameter to decide if center around the empirical mean of the data, default TRUE.}

\item{pbar}{shows the progress bar on max number of iterations, default TRUE.}

\item{phi.start}{set of initial guess for the phases, default NULL.}

\item{standardize}{standardize the matrix used for the inference, default FALSE.}

\item{GTEX_names}{To set to true only if analyzing GTEx data, converts row names from ENSGXXX_GeneName|XX to GeneName. Default FALSE}
}
\value{
A list containing: inferred phases in [0,2*pi], inferred standard deviation of the data, gene parameters gene weights (relevant with two state model),

iteration of stop, inferred standard deviation of data of rhythmic genes, the input matrix, history of the Q function of the EM, genes selected, matrix used for inference.
}
\description{
Infer the circular phase of gene expression samples
}
\examples{
CHIRAL(data_exon)
CHIRAL(data_exon, iterations=50, clockgenes=gene_inf, standardize=TRUE)
}
