% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIRAL.R
\name{cor.c}
\alias{cor.c}
\title{Calculates the circular correlation}
\usage{
cor.c(x, y = NULL)
}
\arguments{
\item{x}{Required vector or matrix. Values MUST be in radiants. If x is a matrix do not provide the parameter y.}

\item{y}{Vector to be provided only if x is a vector too. Must have same length as x. Default NULL}
}
\value{
The value of the circular correlation if x and y are vectors.

A symmetric matrix of dimensions ncol(x) times ncol(x) with entry i,j equal to the circualr correlation between x[,i] and x[,j]
}
\description{
According to A Correlation Coefficient for Circular Data N. I. Fisher and A. J. Lee, https://doi.org/10.2307/2335547
The correlation is calculated either between two vectors or between all the columns of a given matrix. Note that cor.c(matrix) != cor.c(t(matrix)).
}
\examples{
cor.c(matrix)
cor.c(vector_x, vector_y)
}
