% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHIRAL.R
\name{delta.phi}
\alias{delta.phi}
\title{Calculates the best possible shift and direction of two phase distributions to minimize the median absolute error (MAD).}
\usage{
delta.phi(
  phi.0,
  phi,
  period = 2 * pi,
  mode = "forgotten",
  N = 200,
  median_scale = 1
)
}
\arguments{
\item{phi.0}{Required vector of true phases.}

\item{phi}{Required vector of inferred phases. Values MUST be matched to phi.0.}

\item{period}{period of the periodic objects. Default 2*pi}

\item{mode}{Character to decide what the function returns/prints.
If "say" the MAD is printed at screen and not returned, if "return" MAD is only returned, if "no_median" it nor returned not printed, in all other cases it is both returned and printed. Default "forgotten"}

\item{N}{the number of division of the circle where to look for the shift, default 200}

\item{median_scale}{scale the median absolute error when saving/reporting it. Useful in the circadian application to have a median absolute error in hours with phases in radiants.}
}
\value{
Either the inferred phases shifted and or flipped such that the MAD is minimized or a list where "phi" are the best phases and "median" is the value of the MAD
}
\description{
Calculates the best possible shift and direction of two phase distributions to minimize the median absolute error (MAD).
}
\examples{
delta.phi(true.phi, inf.phi, mode="say")
}
